requirements.txt
streamlit==1.31.0
PyPDF2==3.0.1
python-dotenv==1.0.0
openai==1.12.0
langchain==0.1.0
langchain-community==0.0.13
